requirements.txt
pylint
click
google-cloud-language